+++ src/constants.ts
@@ -1,7 +1,10 @@
 export const ACCEPTED_FILE_TYPES = [
-  // 'image/png',
-  // 'image/jpeg',
-  // 'image/jpg',
+  'image/png',
+  'image/jpeg',
+  'image/jpg',
+  'image/webp',
   'video/mp4',
+  'video/webm',
+  'video/quicktime',
   'video/mov'
 ];